Added custom python support and some fixes.
Added the ability to use a custom python command to use. It is stored in /runtime/command.

Changed pypy3.6-v7.1.1-linux64 to pypy3.6-7.1.1-beta-linux_x86_64-portable to fix missing dependency issues on linux systems.

Fixed a crash when starting minecraft or a utility crashing on some systems because of an encoding issue.

Fixed EOL on bat and sh files through .gitattributes. They are now CRLF and LF, respectively.

Unified output format.
Made it so setup behaves almost identically between the sh and bat files.

Python version is now logged instead of being printed to console.

+++ setup.sh
@@ -19,7 +19,13 @@ unpack() {
 }
 
 scriptsonly() {
-	./runtime/bin/python/bin/pypy3 runtime/setuplts.py scriptsonly "$@"
+    echo "> Setting up LTS workspace..."
+    if [ -z "$1" ]; then
+	    ./runtime/bin/python/bin/pypy3 runtime/setuplts.py scriptsonly
+	else
+	    echo "! Running custom python command! Some things may not work correctly!"
+	    "$1" runtime/setuplts.py scriptsonly "$1"
+	fi
 }
 
 start() {
@@ -28,19 +34,17 @@ start() {
     #
 
     # PyPy 3.6.1
-    if [ ! -d runtime/bin/python/ ]; then
+    if [ -z "$1" ] && [ ! -d runtime/bin/python/ ]; then
         echo "> Downloading PyPy 7.1.1 on Python 3.6.1"
 
-        pypy_url=""
         os=$(uname)
-        if [ "$os" == 'Darwin' ]; then
-            pypy_url="https://bitbucket.org/pypy/pypy/downloads/pypy3.6-v7.1.1-osx64.tar.bz2"
-            download $pypy_url runtime/bin/pypy.tar.bz2
-        elif [ "$os" == 'Linux' ]; then
-            pypy_url="https://bitbucket.org/pypy/pypy/downloads/pypy3.6-v7.1.1-linux64.tar.bz2"
-            download $pypy_url runtime/bin/pypy.tar.bz2
+        if [ "$os" == "Darwin" ]; then
+            download "https://bitbucket.org/pypy/pypy/downloads/pypy3.6-v7.1.1-osx64.tar.bz2" runtime/bin/pypy.tar.bz2
+        elif [ "$os" == "Linux" ]; then
+            download "https://bitbucket.org/squeaky/portable-pypy/downloads/pypy3.6-7.1.1-beta-linux_x86_64-portable.tar.bz2" runtime/bin/pypy.tar.bz2
         else
-            echo "! You are not on a supported OS, unfortunately. Sorry about that :("
+            echo "! You are not on a supported OS listed in the autoinstaller, unfortunately. Sorry about that :("
+            echo "! You may try using your own python3 install by using 'c' in setup."
             exit 1
         fi
     fi
@@ -49,13 +53,13 @@ start() {
     # Unzipping natives
     #
 
-    if [ ! -d runtime/bin/python/ ]; then
+    if [ -z "$1" ] && [ ! -d runtime/bin/python/ ]; then
         echo
         echo Unzipping natives
 
         echo "> PyPy 7.1.1 on Python 3.6.1"
         tar -xjf runtime/bin/pypy.tar.bz2 -C runtime/bin/
-        mv runtime/bin/pypy3.6-v7.1.1* runtime/bin/python
+        mv runtime/bin/pypy3.6* runtime/bin/python
         rm runtime/bin/pypy.tar.bz2
     fi
 
@@ -66,31 +70,61 @@ start() {
 	echo
 	echo Setting up LTS workspace...
 
-	./runtime/bin/python/bin/pypy3 runtime/setuplts.py "$@"
+    if [ -z "$1" ]; then
+        ./runtime/bin/python/bin/pypy3 runtime/setuplts.py
+    else
+	    echo "! Running custom python command! Some things may not work correctly!"
+        "$1" runtime/setuplts.py "$1"
+    fi
 
     echo
     echo Finished!
     exit
 }
 
+setpython() {
+    echo "> Enter the path to your desired python install."
+    read -p ": " answer
+
+    start ${answer}
+}
+
+setpythonscriptsonly() {
+    echo "> Enter the path to your desired python install."
+    read -p ": " answer
+
+    scriptsonly ${answer}
+}
+
 if [ -d runtime/bin/python/ ]; then
-    echo "Input 's' if you want to only copy the .sh files."
-    read -p "Are you sure you want to run the setup? [y/N/s]? " answer
+    echo "> Input 's' if you want to only copy the .sh files."
+    echo "> Input 'c' if you want to use a custom python command."
+    echo "> Input 'i' if you want to use a custom python command and only copy the .sh files."
+    echo "> Are you sure you want to run the setup? [y/N/s/c/i]?"
+    read -p ": " answer
 
     case "$answer" in
         [yY])
             start ;;
         [sS])
             scriptsonly ;;
+        [cC])
+            setpython ;;
+        [iI])
+            setpythonscriptsonly ;;
         *)
             exit ;;
     esac
 else
-    read -p "Are you sure you want to run the setup? [y/N]? " answer
+    echo "> Input 'c' if you want to use a custom python command."
+    echo "> Are you sure you want to run the setup? [y/N/c]?"
+    read -p ": " answer
 
     case "$answer" in
         [yY])
             start ;;
+        [cC])
+            setpython ;;
         *)
             exit ;;
     esac
